Move declarations of API structs back to headers
These are not to be used by any implementers directly
at this point. We may define a stable C++ API at some
later point.
