+++ src/sass_values.cpp
@@ -4,84 +4,10 @@
 #include "eval.hpp"
 #include "values.hpp"
 #include "sass/values.h"
+#include "sass_values.hpp"
 
 extern "C" {
   using namespace Sass;
-
-  struct Sass_Unknown {
-    enum Sass_Tag tag;
-  };
-
-  struct Sass_Boolean {
-    enum Sass_Tag tag;
-    bool          value;
-  };
-
-  struct Sass_Number {
-    enum Sass_Tag tag;
-    double        value;
-    char*         unit;
-  };
-
-  struct Sass_Color {
-    enum Sass_Tag tag;
-    double        r;
-    double        g;
-    double        b;
-    double        a;
-  };
-
-  struct Sass_String {
-    enum Sass_Tag tag;
-    bool          quoted;
-    char*         value;
-  };
-
-  struct Sass_List {
-    enum Sass_Tag       tag;
-    enum Sass_Separator separator;
-    size_t              length;
-    // null terminated "array"
-    union Sass_Value**  values;
-  };
-
-  struct Sass_Map {
-    enum Sass_Tag        tag;
-    size_t               length;
-    struct Sass_MapPair* pairs;
-  };
-
-  struct Sass_Null {
-    enum Sass_Tag tag;
-  };
-
-  struct Sass_Error {
-    enum Sass_Tag tag;
-    char*         message;
-  };
-
-  struct Sass_Warning {
-    enum Sass_Tag tag;
-    char*         message;
-  };
-
-  union Sass_Value {
-    struct Sass_Unknown unknown;
-    struct Sass_Boolean boolean;
-    struct Sass_Number  number;
-    struct Sass_Color   color;
-    struct Sass_String  string;
-    struct Sass_List    list;
-    struct Sass_Map     map;
-    struct Sass_Null    null;
-    struct Sass_Error   error;
-    struct Sass_Warning warning;
-  };
-
-  struct Sass_MapPair {
-    union Sass_Value* key;
-    union Sass_Value* value;
-  };
 
   // Return the sass tag for a generic sass value
   enum Sass_Tag ADDCALL sass_value_get_tag(const union Sass_Value* v) { return v->unknown.tag; }
@@ -412,7 +338,7 @@ extern "C" {
 
     // simply pass the error message back to the caller for now
     catch (Error_Invalid& e) { return sass_make_error(e.message.c_str()); }
-    catch (std::bad_alloc& ba) { return sass_make_error("memory exhausted"); }
+    catch (std::bad_alloc&) { return sass_make_error("memory exhausted"); }
     catch (std::exception& e) { return sass_make_error(e.what()); }
     catch (std::string& e) { return sass_make_error(e.c_str()); }
     catch (const char* e) { return sass_make_error(e); }
